backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sse-starlette==2.1.0
deepagents==0.4.12
langchain==0.3.25
langchain-openai==0.3.18
langchain-deepseek==0.1.3
httpx==0.27.0
pydantic==2.9.0
python-dotenv==1.0.1
